cloudFunction/requirements.txt
google-cloud-storage>=1.19.1
google-cloud-texttospeech>=2.2.0